Reverted the last two changes until we can arrive at a solution which also works with MSVC.  The problem was that on Alpha, int64 is simply (long int), thus the format %lld is incorrect (it needs the format %ld).
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@13807 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ libs/csutil/csevent.cpp
@@ -782,16 +782,14 @@ bool csEvent::Print(int32 level)
             || (object->type == attribute::tag_int32)
             || (object->type == attribute::tag_int64))
           {
-            IndentLevel(level); printf (" Value: %lld\n",
-	      (int64)object->Integer);
+            IndentLevel(level); printf (" Value: %lld\n", object->Integer);
           }
           if ((object->type == attribute::tag_uint8)
             || (object->type == attribute::tag_uint16)
             || (object->type == attribute::tag_uint32)
             || (object->type == attribute::tag_uint64))
           {
-            IndentLevel(level); printf (" Value: %lld\n",
-              (uint64)object->Unsigned);
+            IndentLevel(level); printf (" Value: %llu\n", object->Unsigned);
           }
           
           if ((object->type == attribute::tag_float)
